Changes made for 64-bit update.

+++ codebase/apps/didss/src/Fmq2Fmq/Fmq2Fmq.cc
@@ -186,15 +186,14 @@ int Fmq2Fmq::_run ()
     if (gotOne) {
 
       if (_params.debug >= Params::DEBUG_VERBOSE) {
-	cerr << "    Read message, len: " << setw(8) << _inputFmq.getMsgLen()
-	     << ", type: " << setw(8) << _inputFmq.getMsgType()
-	     << ", subtype: " << setw(8) << _inputFmq.getMsgSubtype()
-	     << endl;
+        cerr << "    Read message, len: " << setw(8) << _inputFmq.getMsgLen()
+        << ", type: " << setw(8) << _inputFmq.getMsgType() << ", subtype: " << setw(8)
+        << _inputFmq.getMsgSubtype() << endl;
       }
 
       // add message to write cache
       
-      for (int ii = 0; ii < _outputFmqs_.size(); ii++) {
+      for (size_t ii = 0; ii < _outputFmqs_.size(); ii++) {
         DsFmq &fmq = _outputFmqs[ii];
         if (fmq.isOpen()) {
           PMU_auto_register("write to cache");
@@ -207,7 +206,7 @@ int Fmq2Fmq::_run ()
     
     // write
 
-    for (int ii = 0; ii < _outputFmqs_.size(); ii++) {
+    for (size_t ii = 0; ii < _outputFmqs_.size(); ii++) {
       DsFmq &fmq = _outputFmqs[ii];
       PMU_auto_register("write to output");
       int cacheSize = fmq.getWriteCacheSize();
@@ -252,7 +251,7 @@ void Fmq2Fmq::_openOutputFmqs()
   }
   _prevTimeForOpen = now;
 
-  for (int ii = 0; ii < _outputFmqs_.size(); ii++) {
+  for (size_t ii = 0; ii < _outputFmqs_.size(); ii++) {
     
     if (_params.debug) {
       cerr << "Opening FMQ to URL: " << _params._output_urls[ii] << endl;
